requirements.txt
streamlit
google-generativeai
huggingface_hub